+++ src/components/RadioButton.js
@@ -1,0 +1,29 @@
+import React, { useState } from 'react';
+
+const RadioButton = () => {
+  const [selectedOption, setSelectedOption] = useState('');
+
+  const handleOptionChange = (event) => {
+    setSelectedOption(event.target.value);
+  };
+
+  const radioOptions = ['option1', 'option2', 'option3'];
+
+  return (
+    <div>
+      {radioOptions.map((option) => (
+        <label key={option}>
+          <input
+            type="radio"
+            value={option}
+            checked={selectedOption === option}
+            onChange={handleOptionChange}
+          />
+          {option}
+        </label>
+      ))}
+    </div>
+  );
+};
+
+export default RadioButton;
